style(metal): clear the infer-metal clippy lane
- Remove 3 unnecessary i32->i32 casts and one needless range loop.
- Document the 13 unsafe blocks in CppLfm2Model::build (the workspace
  denies undocumented_unsafe_blocks; the session methods were already
  documented). infer-metal clippy --lib -D warnings is now clean.
- Drop the meaningless const on mlp_compile()'s bool return.

+++ crates/infer-metal/src/slot.rs
@@ -125,9 +125,9 @@ impl MetalSlotState {
             .take()
             .unwrap_or_else(|| new_rows.clone());
         let combined = mlx::concatenate_axis(&[old, new_rows], 0);
-        let len = combined.shape().first().copied().unwrap_or(0) as i32;
+        let len = combined.shape().first().copied().unwrap_or(0);
         self.dflash_target_hidden = Some(if len > 64 {
-            let dim = combined.shape().get(1).copied().unwrap_or(0) as i32;
+            let dim = combined.shape().get(1).copied().unwrap_or(0);
             mlx::slice(&combined, &[len - 64, 0], &[len, dim], &[1, 1])
         } else {
             combined
